frontend/requirements.txt
streamlit
google-cloud-storage
requests